mod.ts のパス誤りを修正
