+++ app/api/e2ee.ts
@@ -1,5 +1,5 @@
 import { Hono } from "hono";
-import { createDB } from "./db.ts";
+import { createDB } from "./db/mod.ts";
 import authRequired from "./utils/auth.ts";
 import { getEnv } from "../shared/config.ts";
 import { rateLimit } from "./utils/rate_limit.ts";
