+++ app/api/routes/activitypub.ts
@@ -1,5 +1,5 @@
 import { Hono } from "hono";
-import { createDB } from "../db.ts";
+import { createDB } from "../db/mod.ts";
 import { getEnv } from "../../shared/config.ts";
 
 import { activityHandlers } from "../activity_handlers.ts";
