+++ app/api/routes/search.ts
@@ -1,5 +1,5 @@
 import { Hono } from "hono";
-import { createDB } from "../db.ts";
+import { createDB } from "../db/mod.ts";
 import { getDomain, resolveActor } from "../utils/activitypub.ts";
 import { getEnv } from "../../shared/config.ts";
 import authRequired from "../utils/auth.ts";
