+++ app/api/routes/session.ts
@@ -1,6 +1,6 @@
 import { Hono } from "hono";
 import { getCookie } from "hono/cookie";
-import { createDB } from "../db.ts";
+import { createDB } from "../db/mod.ts";
 import authRequired from "../utils/auth.ts";
 import { getEnv } from "../../shared/config.ts";
